move 'memoizeWith' tests to separate file
