+++ types/ramda/tsconfig.json
@@ -133,6 +133,7 @@
         "test/max-tests.ts",
         "test/maxBy-tests.ts",
         "test/mean-tests.ts",
-        "test/median-tests.ts"
+        "test/median-tests.ts",
+        "test/memoizeWith-tests.ts"
     ]
 }
